Try to reduce changeset
