+++ hazelcast/src/main/java/com/hazelcast/sql/impl/operation/QueryExecuteOperationFragment.java
@@ -21,10 +21,12 @@ import com.hazelcast.nio.ObjectDataInput;
 import com.hazelcast.nio.ObjectDataOutput;
 import com.hazelcast.nio.serialization.IdentifiedDataSerializable;
 import com.hazelcast.sql.impl.SqlDataSerializerHook;
+import com.hazelcast.sql.impl.plan.node.PlanNode;
 
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Objects;
 import java.util.UUID;
 
 /**
@@ -32,6 +34,7 @@ import java.util.UUID;
  */
 public class QueryExecuteOperationFragment implements IdentifiedDataSerializable {
 
+    private PlanNode node;
     private QueryExecuteOperationFragmentMapping mapping;
     private Collection<UUID> memberIds;
 
@@ -40,11 +43,20 @@ public class QueryExecuteOperationFragment implements IdentifiedDataSerializable
     }
 
     public QueryExecuteOperationFragment(
-            QueryExecuteOperationFragmentMapping mapping,
-            Collection<UUID> memberIds
+        PlanNode node,
+        QueryExecuteOperationFragmentMapping mapping,
+        Collection<UUID> memberIds
     ) {
+        this.node = node;
         this.mapping = mapping;
         this.memberIds = memberIds;
+    }
+
+    /**
+     * @return Operator tree or {@code null} if the fragment should not be executed on the target node.
+     */
+    public PlanNode getNode() {
+        return node;
     }
 
     public QueryExecuteOperationFragmentMapping getMapping() {
@@ -67,7 +79,7 @@ public class QueryExecuteOperationFragment implements IdentifiedDataSerializable
 
     @Override
     public void writeData(ObjectDataOutput out) throws IOException {
-        out.writeObject(null);
+        out.writeObject(node);
         out.writeInt(mapping.getId());
 
         if (memberIds != null) {
@@ -83,6 +95,7 @@ public class QueryExecuteOperationFragment implements IdentifiedDataSerializable
 
     @Override
     public void readData(ObjectDataInput in) throws IOException {
+        node = in.readObject();
         mapping = QueryExecuteOperationFragmentMapping.getById(in.readInt());
 
         int mappedMemberIdsSize = in.readInt();
@@ -96,4 +109,23 @@ public class QueryExecuteOperationFragment implements IdentifiedDataSerializable
         }
     }
 
+    @Override
+    public int hashCode() {
+        return Objects.hash(node, memberIds);
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+
+        QueryExecuteOperationFragment fragment = (QueryExecuteOperationFragment) o;
+
+        return Objects.equals(node, fragment.node) && Objects.equals(memberIds, fragment.memberIds);
+    }
 }
